remove more casting

+++ source/world/entity/Entity.cpp
@@ -186,7 +186,7 @@ void Entity::move(const Vec3& pos)
 
 		AABBVector* cubes = m_pLevel->getCubes(this, AABB(m_hitbox).expand(newPos.x, newPos.y, newPos.z));
 
-		for (int i = 0; i < int(cubes->size()); ++i)
+		for (size_t i = 0; i < cubes->size(); ++i)
 			newPos.y = cubes->at(i).clipYCollide(m_hitbox, newPos.y);
 
 		m_hitbox.move(0.0f, newPos.y, 0.0f);
@@ -195,14 +195,14 @@ void Entity::move(const Vec3& pos)
 
 		bool lastsOnGround = m_bOnGround || (cPosY != newPos.y && cPosY < 0.0);
 
-		for (int i = 0; i < int(cubes->size()); ++i)
+		for (size_t i = 0; i < cubes->size(); ++i)
 			newPos.x = cubes->at(i).clipXCollide(m_hitbox, newPos.x);
 	
 		m_hitbox.move(newPos.x, 0.0f, 0.0f);
 		if (!m_bSlide && cPosX != newPos.x)
 			newPos = Vec3::ZERO;
 
-		for (int i = 0; i < int(cubes->size()); ++i)
+		for (size_t i = 0; i < cubes->size(); ++i)
 			newPos.z = cubes->at(i).clipZCollide(m_hitbox, newPos.z);
 
 		m_hitbox.move(0.0f, 0.0f, newPos.z);
@@ -219,21 +219,21 @@ void Entity::move(const Vec3& pos)
 			m_hitbox = lastHit;
 			cubes = m_pLevel->getCubes(this, AABB(m_hitbox).expand(cPosX, newPos.y, cPosZ));
 
-			for (int i = 0; i < int(cubes->size()); ++i)
+			for (size_t i = 0; i < cubes->size(); ++i)
 				newPos.y = cubes->at(i).clipYCollide(m_hitbox, newPos.y);
 
 			m_hitbox.move(0.0f, newPos.y, 0.0f);
 			if (!m_bSlide && cPosY != newPos.y)
 				newPos = Vec3::ZERO;
 
-			for (int i = 0; i < int(cubes->size()); ++i)
+			for (int i = 0; i < cubes->size(); ++i)
 				newPos.x = cubes->at(i).clipXCollide(m_hitbox, newPos.x);
 
 			m_hitbox.move(newPos.x, 0.0f, 0.0f);
 			if (!m_bSlide && cPosX != newPos.x)
 				newPos = Vec3::ZERO;
 
-			for (int i = 0; i < int(cubes->size()); ++i)
+			for (size_t i = 0; i < cubes->size(); ++i)
 				newPos.z = cubes->at(i).clipZCollide(m_hitbox, newPos.z);
 
 			m_hitbox.move(0.0f, 0.0f, newPos.z);
